+++ src/main/java/org/hv/biscuits/annotation/Affairs.java
@@ -13,5 +13,17 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 public @interface Affairs {
+    /**
+     * 是否开启事务，默认开启
+     *
+     * @return 是否开启事务
+     */
     boolean on() default true;
+
+    /**
+     * 要建立的数据库会话名称，默认为空字符时则取通过其他方式获取sessionName
+     *
+     * @return 要建立的数据库会话名称
+     */
+    String sessionName() default "";
 }