+++ src/main/java/org/hv/biscuits/repository/AbstractRepository.java
@@ -1,5 +1,6 @@
 package org.hv.biscuits.repository;
 
+import org.hv.biscuits.core.session.ActiveSessionCenter;
 import org.hv.pocket.session.Session;
 
 /**
@@ -7,17 +8,7 @@ import org.hv.pocket.session.Session;
  */
 public abstract class AbstractRepository {
 
-    private final ThreadLocal<Session> sessionLocal = new ThreadLocal<>();
-
-    public void injectSession(Session session) {
-        this.sessionLocal.set(session);
-    }
-
-    public void pourSession() {
-        this.sessionLocal.remove();
-    }
-
     public Session getSession() {
-        return this.sessionLocal.get();
+        return ActiveSessionCenter.getCurrentSession();
     }
 }